CRUD COMPLETO
Falta agregar los filtros

+++ go/model/Parada.go
@@ -1,6 +1,6 @@
 package model
 
 type Parada struct {
-	Ciudad       string  `bson:"ciudad"`
-	KmRecorridos float64 `bson:"kmRecorridos"`
+	Ciudad       string `bson:"ciudad"`
+	KmRecorridos int    `bson:"kmRecorridos"`
 }
